+++ CHANGELOG.md
@@ -14,6 +14,8 @@ to a section with the version name.
   * Add error handling for rate limits.
 * Validate buyer address details on submitting `Standard Tax Invoice`.
 * Fix error on invoice creation showing validation message
+* Support all ZATCA discount reasons.
+* Add `ZATCA Discount Reason` field to `Sales Invoice` and `Sales Invoice Item`.
 
 ## 0.50.3